Add comm=ofi large AM capability.
As originally coded, comm=ofi was limited to doing AMs for which the
entire request would fit into a (fairly large) maximum.  But there is no
inherent limit on the size of a request passed in to the comm layer, and
the target-side received request buffers must be of fixed size, so we
also need to be able to support so-called "large" AMs, where only the
header is sent in the initial request, and the target side retrieves the
rest of the arguments using a GET.  Here, implement such large AMs and,
having done so, reduce the size of a non-large AM to 1kb plus the size
of the header.  We have previously established that our AMs tend to run
in the 500-800 byte range, so even with this reduction the majority of
them can still be done with just a single transaction.