Support older versions of fish for setchplenv*.fish
Updates how paths are passed to fixpath for fish.

fish 2.* and 3.* differ in how they print: \\"$PATH\\"

fish 2 would print with space-delimiters and fish 3 would print with
colon-delimiters.

We were relying on this output in the new fixpath.py interface, but it
led to failures on systems with older version of fish. We now rely on
the output of: \" $PATH \" so support both fish 2 and 3.