Add comm=ofi support for an endpoint-per-core model.
The ofi comm layer was written to use a single scalable transmit
endpoint per node and multiple transmit contexts, ideally enough for
each worker thread and the AM handler to have its own.  Unfortunately,
the verbs provider, or more specifically the RxD utility provider used
to support RDM with the verbs provider, does not support scalable
endpoints.  So here, add support for multiple transmit endpoints, again
ideally enough for each worker thread and the AM handler to have its
own.  The decision as to which model to use is automatic, based on
whether the provider says it can support more than one context per
endpoint.

When the scalable model is in use the scalable endpoint is outside the
transmit context table and the table contains tx contexts, as before.
When the multi-endpoint model is in use the endpoint outside the table
is unused and the tx "contexts" in the table are actually endpoints.

Each endpoint needs an address vector, so there is a similar arrangement
with eh AVs: when the scalable model is in use the single AV is stored
in the first transmit context table entry and the other table entries
just alias that.  When the multi-endpoint model is in use each endpoint
has its own AV.

There is a hack in here, marked with the comment "/*hack*/".  That
lengthens the AVs by a factor of 2, because I was seeing internal errors
in the RxD provider due to it not being able to add some kind of
internal pointer to an AV when the AM handler did manual progress on the
RMA endpoint.  I didn't chase down exactly what was going on, but the
internal error was based on not having enough space based on the AV
attribute 'count' value when the AV was created, so I just expanded
that.  It would probably be enough to just do this for the first AV
(which is what the AM and RMA endpoints use).  It would also be good to
figure out exactly what's going on there and fix it in a more principled
way if there is one.  But for the moment I'm leaving it as is.

Note that this still does not run: when the verbs provider is used we
get a segfault when the AM handler does progress on the RMA endpoint.
Based on the traceback it looks like this is because the progress engine
expects that the endpoint should have a CQ associated with it, and all
we have is a counter.  In the interest of keeping the changes separate
I'm going ahead and committing this part now.