Upgrade the progress hinting.
In the past we've forced the use of one or the other progress model,
based on the name of the provider.  That's dumb.  Here, default to
"unspecified" and let the provider tell us which progress model it
prefers.  That done, allow forcing the choice one way or another using
an environment variable, but only when debug support is enabled.