Don't use completion counters if the provider cannot support them.
It appears, based on seeing a number of segfaults having to do with
accessing the CQs of endpoints that only had counters, and then my
noticing the fabric had ofi_info->domain_attr->cntr_cnt==0, that the
verbs provider does not support completion counters.  Thus here, use
counters when that cntr_cnt member indicates that the provider can
support them, and otherwise use CQs.

With this commit, comm=ofi with the verbs provider passes the hello
tests.

While here, add a utility routine to read from a CQ and report errors,
and call it from all the places where we had essentially duplicate code
to do that.

While here, upgrade the verbs-related hack lengthening the AVs to
account for the provider's internal storage requirements so that it's
both large enough to account for all needs as far as I know so far (the
hellos pass, anyway), and documented.

Note: it will probably be easier to look at a diff that ignores white
space changes for this commit, because using the readCQ utility function
changed the indentation for the switch-stmt that differentiates AMs
based on their opcodes but made no other changes to it.