Add support for LAST_ENDPOINT and getsockopt to ZMQ module
Exposes a constant for ZMQ_LAST_ENDPOINT similar to how other of the ZeroMQ
constants have been exposed.

For the getsockopt method, adds an overload for when the return type is
expected to be an int or a string, or when the option is not yet supported.
Distinguishes which one to use via a helper function that looks at the option
provided and returns the appropriate constant from a private enum.  The
various overloads will also throw an IllegalArgument error if a bad argument
is provided.

Adds a runtime include file for the string case.  I'm going to make sure that
works before adding the int case (which supports LINGER).  I'm also going to
check if moving the contents of these functions into a .c file will trigger
issues with the runtime (due to their dependence on ZeroMQ header files, which
may not be present on a user's system)