Restore setting tx_ctx_cnt to the number of transmit contexts.
With a scalable endpoint, the provider's default number of transmit
contexts may be fewer than we want to use.  That being the case, we
need to set that number or we'll get an error when we try to use
what we want.  (An earlier commit to the branch inadvertently
removed this.)