Set ZMQ params to value in zmq.h
We can't set it using the constant itself since those are known at execution
time, not at compile time.  However, it seems as though these numbers are
unlikely to change with later versions of ZMQ.

With that in mind, ensure getsockopt passes the option through instead of in a
way that would require a switch statement to convert from the Chapel params to
the C constants.