Break tie in overload resolution for param and different type
Consider the test coercions-5922.chpl which includes this code:

  1 / (2.0 : real(32))

Before this change, if 2.0 is a real param, this does not resolve,
due to ambiguity between

A. proc /(param a: real(32), param b: real(32))
B. proc /(param a: real(64), param b: real(64))

A. is preferred because the 2nd argument matches the type (real(32)).
B. is preferred because 1 would rather coerce to real(64) than to real(32).

Both of these are "WEAKER" preference. This commit changes the
resolution rule to make params with default size (e.g. 1, 1.0)
prefer the other type (1 to real(64) in this case) with WEAKEST
preference. This breaks the tie in favor of the real(32) version,
which makes sense because it's the one part that the user specified
directly.