Fix signed zeros in complex literal creation
Uses the strategy of default complex number component 0.0
subtracting another value and using the sign from it.

So that 1.0 - 0.0i -> (1.0+0.0i) - (0.0+0.0i) -> 1.0-0.0i
but by the same token, 0.0 - 0.0 -> -0.0

An alternative (that would be harder to implement) would be
to adjust complex param initializations like 1.0 - 0.0i
to construct the number out of the real and imag parts
without first upgrading each to a complex.