Handle potential errors in getsockopt
Make the functions return the error indicator and the Socket methods halt like
setsockopt does.

Since the error codes used by ZMQ are different than common ones, I removed the
qio_error.h include because it is unlikely to be as useful as I thought at the
time.

I would like the ZMQ module to throw errors in the future, but that will take
implementing an Error subclass hierarchy for the ZMQ errors and that seems like
scope creep.  I opened an issue about it, see #12397.