Add test of the new runtime function
Not exposed via Chapel modules just yet because I think that would warrant more
discussion with the group and it is less than a week to feature freeze.  But
the decision should be fairly straight forward, and then we could just import
the Sys module instead of defining the extern function.

Since the hostname will vary from machine to machine, add a preexec that will
get Python's answer and write that to the .good file.