Improve the LLVM back end's ability to find gcc header files (#12410)
Improve the LLVM back end's ability to find gcc header files (#12410)

[reviewed by @ronawho ]

Clang built from gcc depends on gcc's C++ header files. Previously, we were not finding those header files when there was a compiler wrapper in place, such as ccache, or when there was a custom configuration, such as /usr/local/bin/gcc being a symbolic link to gcc in a different directory.

This change adds a more reliable way to find the right directory that will almost always work, and when it fails, we fall back to the old method. (I have not yet been able to find a configuration where the new test does not work.)

Closes #12198