Add a copy of most of the interop tests
These tests are updated to run with --llvm (since most of them compile Chapel
using a .precomp file and are thus unaffected by running start_test with
--llvm).  Most of them work fine, but the array ones have issues.  Specifically,
returning chpl_external_arrays seems to be broken, and chpl_opaque_arrays.

Those I need to investigate.  The --library-makefile tests also fail, but that's
because the code to implement that is currently only being called by the C
backend.  That should be a really quick fix, but I haven't taken the time yet.