Do not promote param and type functions
Before this change a param or type function was considered a valid candidate
for resolution even if it required promotion. If chosen, resolving such
a promoted function would cause one of these errors:

    param function cannot contain a non-param for loop
    illegal assignment of type to value
    iterators may not yield or return types

With this change, such a function will not be considered a valid candidate
at all. If the call ends up unresolved, the compiler will mention
such a function in a "note: candidates are:" message.