Handle []-loops over zip(parallel, serial)
Resolves #11793.
Resolves #12381.
Requires #12419.

* Convert []-loops over `zip(IT1,IT2)` to serial loops
when IT1 provides leader+follower iterators whereas IT2
does not provide a suitable follower, as per #11819.

* Handle forall- and []-loops over tuple expansion expressions
like `(...tup)`.

DETECTING SERIAL-NESS

The compiler now checks availability of followers early.

Before this change, once the compiler found an applicable leader,
it would proceed with the leader+followers setup, without
resolving the leader or looking at the followers.
The followers would be sought and resolved as part of resolving
_toFollowerZip in the followers loop in the body of ForallStmt.
If any of that failed, the tryToken would cover it up and switch
to the serial for-loop.

Now, after finding the leader, the compiler resolves it to find
the type of 'followThis' that will be passed to the follower(s).
Then it checks whether applicable followers exist, by resolving
CallExprs that pass tag=iterKind.follower and followThis.
At this point, follower bodies are not resolved.

If all followers exist, the compiler settles on the L+F direction.
The bodies of the follower iterators are resolved as before -
when resolving the ForallStmt body. A resolution error at this point
becomes fatal. This is the same approach as with a regular function call:
if the user provides a suitable follower iterator, they are responsible
for its correctness.

This last point gets us in trouble when zippering a default-sparse
leader with a default-rectangular follower. The DR follower accepts
any followThis argument in its signature. So the compiler happily
proceeds with L+F. The DR follower, then, fails to resolve and
causes a fatal error because it cannot handle the followThis
that the sparse leader sends to it.

Finally, when the compiler checks whether applicable followers exist
and discovers that one is missing, it switches to implementing it
as a serial loop, relying on #12357.

Essentially the same mechanism, esp. fsIterYieldType and fsGotFollower,
is used both in resolveForallHeader() and when creating promotion
wrappers. Bringing the two flavors closer together is left as future work.

PROMOTION WRAPPERS

Why do we need the above mechanism while creating promotion wrappers?

A while back the compiler would go ahead and create the leader and
follower iterators for a promotion expression in any case. If the
underlying expression did not support leading and/or following,
tryToken would ensure fallback on serial iteration.

Without tryToken, this does not work when a promoted expression is used
as an iterable in a ForallStmt or another promoted expression.

Instead of tryToken, the compiler now relies on the availability of
leaders/followers. Therefore we must not create the leader+follower
for a promoted expression when they wouldn't resolve.

So the compiler checks the underlying expressions for leader/follower
support before creating the leader+follower upon promotion.

This is extra-challenging when zippering a default-sparse leader with a
default-rectangular follower, when the latter was created for a promoted
expression. Because the compiler checks availability of the follower
specifically for **the leader** of that promoted expression.
Upon sparse+DR zippering, the sparse leader produces a different followThis,
and the promoted DR follower cannot handle it, resulting in a resolution
error instead of a fallback onto serial implementation.

The change to SSCA2 is a workaround for this.

TUPLE EXPANSION EXPRESSIONS

For Chapel code like:

```chpl
forall idx in ((...myTuple)) do ....;
```

the AList of iteratedExpressions of the ForallStmt gets a tuple
expansion expression. When expanded, it is replaced with (tuple size)
expressions. So the length of that AList increases, except for 1-tuples.

This happens during resolution. However, the inductionVariables
of the ForallStmt are set up earlier (while parsing). At which time
the size of the tuple may not be known. We used to count the tuple
expansion expression as a single iterable and create a single
induction variable for it.

With this change, we create a single induction variable during parsing
if we notice tuple expansion expression(s) among the iteratedExpressions.
If so, we make the necessary adjustments to inductionVariables
during resolution. Luckily, the followers loop already handled this
situation so needed little adjustment, if any.

In retrospect, we should probably move the setup of inductionVariables
during parsing into resolution / resolveForallHeader().

COMPILER CODE RESTRUCTURING

I moved around + changed signatures of some code in foralls.cpp
wrappers.cpp to handle the new scenarios.

TESTING CHANGES

SSCA2 - `torus_graph_generator_utilities.chpl` - a workaround for #12424
  here the case of zip(SPS array, DR-promoted expression).

`test/types/range/diten/badRange2.good` - the compiler now resolves
  the leader iterator before the serial iterator.

`zippered-serial.chpl` - moved to balance number of tests per directory.