Restore the ability to build the LLVM verifier (#12434)
Restore the ability to build the LLVM verifier (#12434)

[reviewed by @mppf ]

#12102 deleted the LLVM test directories to save a large amount of files and space in our included version of LLVM. It turns out that after that PR, setting CHPL_LLVM_DEVELOPER=1 and rebuilding Chapel causes an abort due to an attempt to access the missing test directories. They are not actually needed, but are only accessed by the build procedure.

Under some conditions, such as during a Debug build, the value of LLVM_INCLUDE_TESTS is not preserved and is set to its default. It is not apparent how this is happening; no reference to LLVM_INCLUDE_TESTS in the code is the culprit. There may be some configurations in which most variables are set to their default.

This change sets the default for LLVM_INCLUDE_TESTS to OFF, since passing -DLLVM_INCLUDE_TESTS=OFF does not always work. This is necessary due to the way we package LLVM, and would not be a problem in their full distribution.