Add an UnorderedCopy that will replace BufferedGets
BufferedGets/buffGet reveals too much implementation detail and makes
implicit comm explicit (and only worked when the destination was local.)
UnorderedCopy/unorderedCopy is better in these regards. I considered
UnorderedAssign, but want to make it more clear that a bitcopy is being
done and a user defined assignment overloads will not be invoked.