Error for "for zip(myIter(tag=leader), ...)"
SUMMARY

* Check for a case for which we do not have good semantics
  and which we most likely do not compile properly today.
  Background:
    https://github.com/chapel-lang/chapel/pull/12357#discussion_r258208652

* Adjust tests in functions/iterators/elliot/isRefIter,
  see details below.

* Simplify the code, including removal of some pattern matching
  and removal of the term "eflopi".

* Introduce isParallelIterator(), for convenience.

DETAILS

This change simplifies the code in implementForallIntents2.cpp and
strengthens checking for a corner case where we could potentially
generate racy code.

That change broke the tests in functions/iterators/elliot/isRefIter,
which reported an error whenever a leader or standalone iterator call
is passed to printIterRefnessWrapper(), which tries to iterate over it.

    // ok
    for i in printIterRefnessWrapper(c.myValIter()) do;

    // "unimplemented" error
    for i in printIterRefnessWrapper(c.myValIter(tag=iterKind.standalone)) do;

I adjust the tests to pass leader/standalone iterator calls to a slight mod
of printIterRefnessWrapper(), which does not iterate over it.

While there, I factor out the helper procs+iters into a separate file.

MORE DETAILS - 1

When we pass a leader or standalone iterator to printIterRefnessWrapper(),
the compiler converts its "for i in myIter" loop to a ForallStmt.
See below for the reason.

A ForallStmt needs to have a CallExpr as its iterable. For example:
```chpl
forall i in someIter() do ....
```

Because a ForallStmt is lowered by inlining that call. It is like a
ForLoop that is lowered by inlining. Except a ForallStmt currently
does not have an alternative implementation that works without inlining.

A ForallStmt over a formal argument falls into this unimplemented case.
Because there is no readily-available CallExpr to inline.

So the compiler generates an error saying that. This happens in the
`for i in myIter` case when myIter is passed a parallel iterator.

MORE DETAILS - 2

Why convert a for-loop over a parallel iterator to a forall-loop?

This is needed to ensure no data races. For example:

    forall .... with (+ reduce sum) {
      writeln("sum=", sum);
      for idx in myIter(tag=iterKind.standalone) do
	sum += idx;
    }

When compiling a forall-loop, each task gets its own shadow variable
for 'sum'. The writeln() implicitly accesses the shadow variable for its task.

Cf. when compiling a for-loop, all iterations access the same 'sum'
variable. In the above example, that would be the shadow variable
accessed by the writeln().

If myIter(tag=iterKind.standalone) generates concurrent tasks,
they will race for that shadow variable.

To avoid that, the compiler checks each for-loop for a parallel
iterator. When found, it converts the AST representation from ForLoop
to an equivalent ForallStmt.