Improve finding include and lib files for comm=ofi and mpirun4ofi.
There are two changes here in support of comm=ofi, the mpirun4ofi
launcher and, by extension, the MPI-based out-of-band support for
comm=ofi.  The smaller one is that we had been using separate
environment variables for the locations of the MPI include and lib
dirs.  Here, switch to just a single MPI_DIR env var for that, and
assume it will have include and lib subdirs.

The larger change is to deal with some variability in paths for
libfabric and MPI libraries.  Specifically, we've encountered
systems where libfabric.{a,so} is in a subdir called .../lib64
instead of just .../lib, and also systems where the MPI is MPICH
instead of OpenMPI and the library is libmpich.{a,so} instead of
libmpi.{a,so}.  So, adjust the user program link-time command line
arguments to deal with that.