Squash "In routine..." reporting for internal function names
For calls to compilerError(), compilerWarning() and the like, we
generate a prefix that says "In function..." or "In iterator..."
and then names the function.  The compiler does this even if the
function in question is internal, which seems not very helpful
and potentially confusing/noisy.  This PR squashes such cases for
internal functions starting with 'chpl_' when --devel is not thrown.