Changes reflecting review feedback.
Use make's wildcard function instead of shell syntax to look for the
.../lib64 and .../lib directories and the libmpich.* files.  Assume that
if the MPI library is not named libmpich.* then it is named libmpi.*;
don't bother looking specifically for the latter.  As a helpful measure,
throw -lmpi even if we don't find an MPI library directory: it will show
up in --print-commands output demonstrating that we knew we needed it,
and the linker will have a chance to say "couldn't find -lmpi".  (We do
the same for -lfabric even if we can't find the lib dir for that.)