Update test based on lack of identity()
With the new stricter "state type == result type" check, this test
started failing because the query of identity could not be performed.
I added an identity which showed that the state types does not match
the result type for this reduction which disables the parallel
implementation.  So then I updated the error message again to match
what it should be.