First good draft of parallel Block-distributed scan
This implements a parallel 1D scan on block-distributed arrays by
refactoring the doiScan on default rectangular arrays so that we can
call its two passes as helper functions, chpl__preScan and
chpl__postScan (methods on default rectangular arrays).

For the Block distribution, it then combines this with a serial
(currently) inter-locale scan of the per-locale contributions between
the two passes in order that each locale can only touch its complete
data twice.