Reduce memory leaks due to reduce expressions
This should revert the memleaks increase due to #12131. See #12125.

* Reduce the amount of temporaries added by lowerPrimReduce() and
insertFinalGenerate().

* Which, together with adding FLAG_EXPR_TEMP, takes care of the memleaks,
however causes valgrind failure in two tests, see #12497.

* Remove FLAG_EXPR_TEMP, effectively do not add it, in the special case
when it causes valgrind failure. See workaroundForReduceIntoDetupleDecl()
and #12497.

* Add a minimal-reproducer test for #12497. While it is not essential,
I am adding it to facilitate resolution of #12497 and just in case
the other two failing tests are modified and do not exhibit the problematic
pattern any longer.

While there: change lowerPrimReduce() to return the result instead of
assigning it into a ref formal.