Fix LLVM optimizer specialization on Cray systems (#12530)
Fix LLVM optimizer specialization on Cray systems (#12530)

[reviewed by @ronawho ]

The Clang CPU flag was not being computed correctly when a PrgEnv module was present. Sometimes the PrgEnv name for a CPU is identical to the Clang CPU name, which is why this was not detected earlier. This change computes the Clang CPU name when Clang specialization is being done, and retains the PrgEnv CPU name otherwise.