Remove more orphan AST nodes
Resolves #12031.

The compiler now removes some ASTs that reference
already-removed-from-AST types. Tests:

    test/expressions/vass/eval-of-arg-of-.type.chpl
    test/arrays/deitz/runtime_types/test_array_type4.chpl

The main focus is the situation when a "type" BlockStmt that is removed
contains ASTs for a loop expression. The removal of the BlockStmt includes
removal of the corresponding _iteratorRecord+_iteratorClass types.

Background: a "type" BlockStmt is typically created for a Chapel expression
`EXPR.type`. The block is removed in removeTypeBlocks(), assuming that
the resulting type is non-runtime.

The ref types for these are left around, and so are iteratorIndexType/Zip
and autoCopy FnSymbols whose formals reference them. If the type block
is at the module level, its temps are moved to the module scope.
If so, a no-longer-used temp for the _iteratorClass is left around, too.

To handle this situation, the compiler now does two conceptually different
things:

* `removeDeadTypeStuff()` - removes ref types **specifically** for those
  types that are being removed because their defs are in the type block.

* `removeSymbolsWithRemovedTypes()` - removes vars and functions with
  already-removed types, **regardless** of the cause of removal.

We could apply the latter "regardless of cause" approach more aggressively
to do the following:

  * removals presently done by `removeDeadTypeStuff()`
  * remove functions whose return type is not inTree()
  * remove VarSymbols whose firstSymExpr() is NULL
  * remove **all** ref types whose base type is not inTree()
  * remove _iteratorRecord, _iteratorClass, and the various
    iterator-related functions for unused and/or pruned iterators

I did not pursue this route due to us close to code freeze.

Separately, the compiler now checks whether the type produced
by a type block is defined within that type block. If so,
complain bitterly to the user. Issue a "not supported" error, that is.
Test:

    test/functions/iterators/errors/iterator-.type.chpl

Code organization: I considered putting removeDeadTypeStuff()
in functionResolution.cpp, as it is close in spirit to other
cleanup code there. Instead, I chose to move it off into LoopExpr.cpp
to reduce the size of functionResolution.cpp. We should probably
move the cleanup code altogether out of functionResolution.cpp
into a separate file under resolution/.

While there:
  * simplify some code in functionResolution.cpp by using for_alive_in_Vec
  * expand Type::inTree() to remove the check for rootModule (never fires)
  * simplify removeRandomPrimitive() because it is used only with primitives