Fix checking for virtual iterators
Resolves #12123.

This restores the compiler's ability to issue the error
"virtual parallel iterators are not yet supported"
when a dynamically-dispatched iterator is used in a forall loop.

This error reporting was based on checking `isVirtualIterator()`,
which is passed the iterator function's return type as the argument.
In #12131, the iterator functions used in forall loops became
subject to the return-by-ref transformation. As a result, their
return type became `void` and `isVirtualIterator()` was no longer
detecting virtuality.

This modifies `isVirtualIterator()` to accept the iterator function
itself and to get the formerly-return type from its IteratorInfo.
Then proceed analyzing that type as before.
When the function is an iterator forwarder, get the formerly-return
type from its ret-arg formal.

This restores the pre-#12123 contents of the .bad and .future files
for the tests in `functions/iterators/elliot/dynamicDispatch` .

While there, adjust .bad for a change in assertion,
which missed in #12538.

Trivial, not reviewed.