Fix a memory error
Once I closed the obvious link, I ran into an issue where a loop
was trying to traverse the list while freeing its elements.  Due
to the way the these() iterator was written, it referred to the
freed element after yielding it.  By storing the 'next' field in
a temporary variable, I was able to work around this.