Changes reflecting review feedback.
Revert the change to the InfiniBand doc, so that we don't seem to be
implying users should consider switching to comm=ofi, which after all
isn't fully ready yet.  While here, though, remove the spurious "The" at
the start of the original sentence.

Remove idiosyncratic line breaks in chplenv.rst.