Fix issues related to --target-cpu
In doing a final test tonight, I was finding that
test/chplenv/autocomplete/autocomplete.chpl was failing due to the new
--target-cpu flag.  It turns out that the reason is that the scripts
were failing to strip out the argument and description for the flag
due to the fact that the description contained a `--` pattern.  This
showed up in Michael's original update to the file today, but was
removed after the fact (presumably manually) and before merging the
PR.  However, doing so broke the autocomplete test which runs the same
script and got the same bad output.

I'm not enough of a sed expert to want to try and fix its pattern
matching tonight, so instead I reworded the description of the flag to
avoid '--' in its description which I think still scans fine.  I also
fixed the alphabetization of the flag (which is not strictly required,
but in this case seemed like a mistake, or at least didn't have an
obvious rationale).  This in turn required updating the
chpl-completion.bash file and userhelp.good to match the new ordering.