Fix cross-compilation for GMP builds
This fixes GMP build issues when cross-compiling (for example when
building on a Cray with the craype-broadwell CPU targeting module on a
sandybridge login node.)

GMP has a `CC_FOR_BUILD` parameter. From the docs:
"""
Some build-time programs are compiled and run to generate host-specific
data tables. ‘CC_FOR_BUILD’ is the compiler used for this. It doesn’t
need to be in any particular ABI or mode, it merely needs to generate
executables that can run.
"""

We weren't aware of this before, but ran into some issues when trying to
cross-compile that exposed this. Since GMP will compile/run programs on
the host/build/login node, set `CC_FOR_BUILD` to `CHPL_HOST_COMPILER`.