Start on generalized support for oversubscribed execution.
Up until now we've been dealing with oversubscribed execution in an
ad-hoc way.  For example, when we expect to need to run oversubscribed
we set CHPL_QTHREAD_ENABLE_OVERSUBSCRIPTION when building Qthreads,
and then we set QT_AFFINITY and QT_SPINCOUNT when actually running
that way.  (Mostly this is intended just for our own testing.  We do
document CHPL_QTHREAD_ENABLE_OVERSUBSCRIPTION for users, but not the
other two.)

Here, begin supporting oversubscribed execution in a more general way,
starting with the ofi comm layer since we've discovered some settings
that improve its performance in oversubscribed situations on Mac OS X.
Define a new boolean environment variable CHPL_RT_OVERSUBSCRIBED.  If
it is true at launch time, for comm=ofi set the environment variable
FI_SOCKETS_PE_WAITTIME=0 to reduce the overhead of automatic progress
in the sockets provider.  Also, if the mpirun4ofi Chapel launcher is
being used then set the OpenMPI MCA parameter mpi_yield_when_idle=1 to
cause MPI to yield the processor when waiting for MPI communication,
because currently using the mpirun4ofi launcher also causes us to use
comm=ofi's MPI-based out-of-band support, which means MPI is present,
though idle, in all comm=ofi Chapel programs launched that way.  (This
parameter is documented as being set automatically when MPI is in an
oversubscribed situation, but based on test runs it seems that our
throwing '-map-by node:oversubscribed' to mpirun is not sufficient to
cause that to happen.)