Update version of docs/master to be 1.20
This should really have gone in with my previous commit to
versionButton.php...  Tony and I were still getting the dance
w.r.t. these files correct in the last release and I forgot about the
dependence between the version that was embedded in the docs by this
configuration file.