Update test for more pre-existing memory regions on CLE.
The overflow-mem-region-table test has started failing.  The reason
turns out to be that the ugni comm layer is finding more pre-existing
memory regions when it starts up, enough that it is running out of
memory region table entries during startup rather than during user
program execution, and thus producing a different message than the
one we expect it to.  Here, adjust the test configuration to allow
for more memory regions, enough to get through startup, but also to
create more so that we're still guaranteed not to have enough and get
the desired error mesage.  Also, relax the prediff file's filtering
so that it will be easier to see what's happened the next time this
occurs.