Add a test that reads HDF5 files into a distributed array
[reviewed by @mppf]

This work was done with ideas from @mppf whose thoughts dramatically
simplified the direction and clarity of the code.

This code reads one file/dataset per locale into a distributed array. The
datasets in each file do not need to be the same size and do not need to be
completely local to the locale that is reading them.

The .h5 files created by this test each contain datasets with 10,000 + a
small random number of c_int elements. When read properly the array will
contain 1,2,3,... The test deletes the .h5 files before it finishes.

The total number of gets/puts done by the reader function can be counted and
printed with --printCommCounts=true.