Simplify the borrowing of the slicing domain
Previously, we were trying to be clever about using the existing
domain in some cases, but not in others (I'd forgotten about this
logic).  This simplifies things by just always using the slicing
domain.

This code fails for about 30 tests because we try to register
the array view with the domain and that results in an unexpected
type internal error.  Are array views not sufficiently array-like?

Interestingly, the array views don't need to be reallocated when
a domain changes, but they do need to keep a domain alive if it
would otherwise go out of scope.