Update test for more pre-existing memory regions on CLE, again.
The overflow-mem-region-table test is still failing in some configs,
namely a few where we still have more initial memory regions than the
specified table size and so get an unexpected internal error instead of
the expected user-level error.  Make the table size larger to allow for
those situations, and make the demand parameter larger, too, so as to
ensure we still run into the limit even when those extra initial regions
aren't present.