Improve comm=ugni SIGBUS handler.
The SIGBUS handler in the ugni comm layer was sloppy.  Harden it by
adding an entry gate so that no more than one thread can use it at a
time, and by only calling functions that are safe in signal handlers.

This seems to get rid of the malloc() gripes about arena corruption that
were present in the non-prediff'ed output.  It also seems to get rid of
the timeouts which began recently and which we believe were the result
of trying to produce (extremely large) core dumps due to secondary
errors, perhaps SIGBUS when that was masked inside the handler.