Deprecate the -senableParScan config param
The 'enableParScan' config param was added because the parallel scan
feature was added so late in the release cycle and we hadn't had much
time to live with it to gain confidence in it yet.  Here, I'm
deprecating the flag and enabling the parallel implementations (when
available) since we're at the start of a release cycle.