Allow _exit() calls in the runtime.
Remove _exit() from the list of functions that mustn't be called from
the runtime.  It's easy to make a case for not calling exit() (no
leading underbar), because that's the normal exit function and one might
easily forget about chpl_exit_*() and use exit() instead.  But _exit()
(with an underbar) is specifically for cases when we need to halt right
now, without doing anything else, for example from a signal handler in
which there's no hope of rescuing the situation.  The only alternative
in that situation is abort(), but it has other effects which may not
desired, notably a core dump.