Throw errors instead of halting (with a new ZMQError class)
Eventually we'd like there to be some sort of hierarchy used for ZMQ errors.  It
might be merged with SystemError in a reasonable way, or it might end up being
its own subclass.

Note that this does not update previous methods that halt instead of throwing
errors.  That is later work as well.