Add tests for all bits and change input to octal
This add tests for all the bits (rwx) for users, groups and others.
It also has tests for the setuid, setgid and sticky bits.
The input for the tests were in decimal, meaning conversion was required to understand the mode.
I changed all of them to octal, making the mode directly understandable.
There are a total of 19 tests.
All of them ran successfully using the `start_test` command.
Finally, I created a separate folder named `chmod` under `FileSystem` to move away from having folders with people's names.
If this is merged, the `chmod` folder under `lydia` will not be needed.